seed-server: make `IPv4` the first case
Previously we checked for `IP_FAMILIES==IPv4` so we shouldn't switch to checking `IP_FAMILIES==IPv6` in if/else, which is different.
This commit goes back to the previous `IP_FAMILIES=IPv4` check but keeps the inverted logic of first checking the IP family and then the HA config.

It also simplifies calculation of the first three VPN network octets.