Always listen on both IPv4 and IPv6 sockets on server side.
To simplify the configuration by removing unnecessary options and make
it easier to support dual-stack configurations on server side with
single-stack on client side, the vpn server is configured to always
listen on both IPv4 as well as IPv6 sockets. Unfortunately, this is
achieved via the unintuitive `tcp6-server` option.
Looking at it via tools like `netstat` or `ss`, it looks like there is
only an IPv6 socket, but you can also address it via IPv4.